Moved AoA into common to avoid cyclic dependancies
