--- ArticlesOfAssociation/Team4AoA.go
@@ -1,1 +1,0 @@
-package aoa
