--- ArticlesOfAssociation/Team5AoA.go
@@ -1,1 +1,0 @@
-package aoa
