--- ArticlesOfAssociation/Team6AoA.go
@@ -1,1 +1,0 @@
-package aoa
