+++ common/ArticlesOfAssociation.go
@@ -1,4 +1,4 @@
-package aoa
+package common
 
 import "github.com/google/uuid"
