+++ common/IExtendedAgent.go
@@ -1,7 +1,6 @@
 package common
 
 import (
-	aoa "SOMAS_Extended/ArticlesOfAssociation"
 	"github.com/MattSScott/basePlatformSOMAS/v2/pkg/agent"
 	"github.com/google/uuid"
 )
@@ -48,6 +47,6 @@ type IExtendedAgent interface {
 	LogSelfInfo()
 	GetAoARanking() []int
 	SetAoARanking(Preferences []int)
-	GetContributionAuditVote() aoa.Vote
-	GetWithdrawalAuditVote() aoa.Vote
+	GetContributionAuditVote() Vote
+	GetWithdrawalAuditVote() Vote
 }
