+++ common/Team1AoA.go
@@ -1,4 +1,4 @@
-package aoa
+package common
 
 import (
 	"container/list"
