+++ common/team.go
@@ -2,7 +2,6 @@ package common
 
 import (
 	// TODO: should it be structured this way?
-	aoa "SOMAS_Extended/ArticlesOfAssociation"
 
 	"github.com/google/uuid"
 )
@@ -10,7 +9,7 @@ import (
 type Team struct {
 	TeamID  uuid.UUID
 	Agents  []uuid.UUID
-	TeamAoA aoa.IArticlesOfAssociation
+	TeamAoA IArticlesOfAssociation
 
 	commonPool int
 }
@@ -25,7 +24,7 @@ func (team *Team) SetCommonPool(amount int) {
 
 // constructor: NewTeam creates a new Team with a unique TeamID and initializes other fields as blank.
 func NewTeam(teamID uuid.UUID) *Team {
-	teamAoA := aoa.CreateFixedAoA()
+	teamAoA := CreateFixedAoA()
 	return &Team{
 		TeamID:     teamID,        // Generate a unique TeamID
 		commonPool: 0,             // Initialize commonPool to 0
